requirements.txt
lxml
numpy
tensorflow==1.15
pillow
pandas

